just on production

+++ webhook.php
@@ -3,7 +3,7 @@
 		'res'=>'nok'
 	);
 
-	if (@$_POST['payload'] ) {
+	if (@$_POST['payload'] && $_POST['payload']['ref'] == 'refs/heads/production') {
 		$out = shell_exec( 'cd /usr/share/drupal7/legacy/eamt2017/ && git reset --hard HEAD && git pull origin production 2>&1' );
 		$res = array(
 			'res'=>'ok',
